Refactor CLI profile helpers into SRP modules
