+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,2 @@
+pub(super) mod bundle_inspection;
+pub(super) mod profiles;
